requirements.txt
streamlit
python-docx
PyPDF2
spacy
pdfplumber
